requirements.txt
evdev==1.9.2
Kivy==2.3.1
KivyMD==1.2.0
pyserial>=3.5,<4
